separating AnecdoteApp into modules
