+++ redux-anecdotes/src/App.jsx
@@ -1,45 +1,12 @@
-import { useSelector, useDispatch } from 'react-redux'
-import { createAnecdote, vote} from './reducers/anecdoteReducer'
+import AnecdoteForm from './components/AnecdoteForm.jsx'
+import AnecdoteList from './components/AnecdoteList.jsx'
 
 const App = () => {
-
-  const dispatch = useDispatch()
-  const anecdotes = useSelector(state => state)
-
-  const addAnecdote = (event) => {
-    event.preventDefault()
-    const content = event.target.anecdote.value
-    event.target.anecdote.value = ''
-
-    dispatch(createAnecdote(content))
-  }
-
-  const voteAnecdote = (id) => {
-    dispatch(vote(id))
-  }
-
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
-          <div>
-            {anecdote.content}
-          </div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => voteAnecdote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      )}
-      <h2>create new</h2>
-      <form onSubmit={addAnecdote}>
-        <div>
-          <input name="anecdote" />
-        </div>
-        <button type="submit">create</button>
-      </form>
+      <AnecdoteList />
+      <AnecdoteForm />
     </div>
   )
 }
