+++ redux-anecdotes/src/components/AnecdoteForm.jsx
@@ -1,0 +1,25 @@
+import { useDispatch } from 'react-redux'
+import { createAnecdote } from '../reducers/anecdoteReducer'
+const AnecdoteForm = () => {
+    const dispatch = useDispatch()
+    const addAnecdote = (event) => {
+        event.preventDefault()
+        const content = event.target.anecdote.value
+        event.target.anecdote.value = ''
+    
+        dispatch(createAnecdote(content))
+    }
+    return (
+        <>
+        <h2>create new</h2>
+        <form onSubmit={addAnecdote}>
+            <div>
+            <input name="anecdote" />
+            </div>
+            <button type="submit">create</button>
+        </form>
+        </>
+      )
+}
+
+export default AnecdoteForm
