+++ redux-anecdotes/src/components/AnecdoteList.jsx
@@ -1,0 +1,27 @@
+import { useSelector, useDispatch } from 'react-redux'
+import { vote } from '../reducers/anecdoteReducer'
+const AnecdoteList = () => {
+    const dispatch = useDispatch()
+    const anecdotes = useSelector(state => state)
+
+    const voteAnecdote = (id) => {
+        dispatch(vote(id))
+      }
+    return (
+        <>
+            {anecdotes.map(anecdote =>
+        <div key={anecdote.id}>
+          <div>
+            {anecdote.content}
+          </div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => voteAnecdote(anecdote.id)}>vote</button>
+          </div>
+        </div>
+      )}
+        </>
+    )
+}
+
+export default AnecdoteList
